% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crown.R
\docType{package}
\name{crown-package}
\alias{crown-package}
\title{Cluster-Randomized Trial Analysis with Outcomes Weighted for Nonresponse}
\description{
G-formula, IPW, and AIPW estimators combine trial outcomes with auxiliary
covariates using logistic regression or cross-fitted machine learning.
}
\author{
\strong{Maintainer}: Zhitan He \email{zhitanhe@unc.edu}

Authors:
\itemize{
  \item Brian Richardson \email{richabr@wfu.edu}
}

}
\keyword{internal}
