% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gfmla_fit.R
\name{gfmla_fit_naive}
\alias{gfmla_fit_naive}
\title{Naive g-formula estimator}
\usage{
gfmla_fit_naive(dat, mu_fmla)
}
\arguments{
\item{dat}{data frame containing the following columns:
\itemize{
\item \code{C}: a binary indicator for whether the outcome is censored (\code{C}=1) or
not (\code{C}=0)
\item \code{A}: a binary exposure
\item \code{Y}: a binary outcome
\item other covariates specified in \code{mu_fmla}
\item \code{wt}: survey weights
}}

\item{mu_fmla}{a formula for the outcome regression model using variables in
\code{dat}}
}
\value{
a list containing the following:
\itemize{
\item \code{eta_hat}: a numeric vector estimated mean potential outcomes eta(0)
and eta(1)
\item \code{eta_hat_cov}: a numeric matrix, estimated covariance of \code{eta_hat}
\item \code{outcome_reg}: a list, results of outcome regression model
}
}
\description{
estimate eta(0) and eta(1) using a naive g-formula estimator, i.e., ignoring
nonresponse
}
