% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipw_fit.R
\name{ipw_fit}
\alias{ipw_fit}
\title{Crown IPW estimator}
\usage{
ipw_fit(dat, pi_fmla)
}
\arguments{
\item{dat}{data frame containing the following columns:
\itemize{
\item \code{S}: a binary indicator for whether the observation belongs to the trial
data (\code{S}=1) the auxiliary data (\code{S}=0)
\item \code{R}: a binary indicator for whether the observation is a responder
(\code{R}=1) or not (\code{R}=0)
\item \code{C}: a binary indicator for whether the outcome is censored (\code{C}=1) or
not (\code{C}=0)
\item \code{A}: a binary exposure
\item \code{Y}: a binary outcome
\item other covariates specified in \code{mu_fmla}
\item \code{wt}: survey weights

}}

\item{pi_fmla}{a formula for the propensity score regression model using
variables in \code{dat}}
}
\value{
a list containing the following:
\itemize{
\item \code{eta_hat}: a numeric vector estimated mean potential outcomes eta(0)
and eta(1)
\item \code{eta_hat_cov}: a numeric matrix, estimated covariance of \code{eta_hat}
\item \code{Q_reg_0}: a list, results of Q0 membership regression model
\item \code{Q_reg_1}: a list, results of Q1 membership regression model
\item \code{dat}: a data frame including a column for propensity score weights
}
}
\description{
estimate eta(0) and eta(1) using a crown IPW estimator, i.e., accounting for
nonresponse
}
