% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonpar_est.R
\name{nonpar_est}
\alias{nonpar_est}
\title{Nonparametric estimation wrapper}
\usage{
nonpar_est(x, y, method, arguments = NULL, wts = NULL)
}
\arguments{
\item{x}{a numeric matrix of covariates}

\item{y}{a binary vector of outcomes}

\item{method}{either "xgboost" or "SuperLearner", method of nonparametric
estimation}

\item{arguments}{an optional list with arguments to pass to \code{nonpar_est}. If
\code{method} is "SuperLearner", \code{arguments} must contain \code{SL.library} and
\code{cvControl}}

\item{wts}{an optional numeric vector of observation weights}
}
\value{
a fitted model
}
\description{
estimate the conditional distribution of a binary outcome using either
xgboost or SuperLearner
}
