% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonpar_est.R
\name{nonpar_pred}
\alias{nonpar_pred}
\title{Nonparametric prediction wrapper}
\usage{
nonpar_pred(mod, newdata, method)
}
\arguments{
\item{mod}{a fitted model object, output of \code{nonpar_est}}

\item{newdata}{a numeric matrix, predictors from new data}

\item{method}{either "xgboost" or "SuperLearner", method of nonparametric
estimation}
}
\value{
a numeric vector, predicted probabilities
}
\description{
predict probabilites using a fitted model from xgboost or SuperLearner
}
