% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dml_fit.R
\name{dml_fit}
\alias{dml_fit}
\title{Crown DML estimator}
\usage{
dml_fit(dat, mu_covariates, pi_covariates, K, method, arguments = NULL)
}
\arguments{
\item{dat}{data frame containing the following columns:
\itemize{
\item \code{S}: a binary indicator for whether the observation belongs to the trial
data (\code{S}=1) the auxiliary data (\code{S}=0)
\item \code{R}: a binary indicator for whether the observation is a responder
(\code{R}=1) or not (\code{R}=0)
\item \code{C}: a binary indicator for whether the outcome is censored (\code{C}=1) or
not (\code{C}=0)
\item \code{A}: a binary exposure
\item \code{Y}: a binary outcome
\item other covariates specified in \code{mu_fmla}
\item \code{wt}: survey weights
}}

\item{mu_covariates}{a formula for the outcome regression model using
variables in \code{dat}}

\item{pi_covariates}{a formula for the propensity score regression model
using variables in \code{dat}}

\item{K}{a positive integer number of folds for sample splitting}

\item{method}{a character string, method of nonparametric estimator}

\item{arguments}{an optional list with arguments to pass to \code{nonpar_est}}
}
\value{
a list containing the following:
\itemize{
\item \code{eta_hat}: a numeric vector estimated mean potential outcomes eta(0)
and eta(1)
\item \code{eta_hat_cov}: a numeric matrix, estimated covariance of \code{eta_hat}
\item \code{dat}: a data frame with estimated propensity score weights
}
}
\description{
estimate eta(0) and eta(1) using a crown DML estimator, i.e., accounting for
nonresponse, with nonparametric sample splitting
}
